\name{coefficientalpha-package}
\alias{coefficientalpha-package}
\alias{coefficientalpha}
\docType{package}
\title{
Robust Cronbach's alpha with missing and non-normal data
}
\description{
An R package to calculate coefficient alpha with missing data and non-normal data. Robust standard error and confidence interval are also provided.
}
\details{
\tabular{ll}{
Package: \tab coefficientalpha\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-01-22\cr
License: \tab GPL\cr
}
}
\author{
Zhiyong Zhang and Ke-Hai Yuan
Maintainer: Zhiyong Zhang <zzhang4@nd.edu>
}
\references{
Zhang, Z. & Yuan, K.-H. (2013). Robust Cronbach's alpha.
}
\keyword{robust}
\keyword{alpha}
\keyword{Cronbach}

