\name{coefficientalpha-package}
\alias{coefficientalpha-package}
\alias{coefficientalpha}
\docType{package}
\title{
 Robust Coefficient Alpha and Omega with Missing and Non-Normal Data
}
\description{
Cronbach's alpha and McDonald's omega are widely used reliability or internal consistency measures in social, behavioral and education sciences. Alpha is reported in nearly every study that involves measuring a construct through multiple test items. The package 'coefficientalpha' calculates coefficient alpha and coefficient omega with missing data and non-normal data. Robust standard errors and confidence intervals are also provided. A test is also available to test the tau-equivalent and homogeneous assumptions. Since Version 0.5, the bootstrap confidence intervals were added.
}
\details{
\tabular{ll}{
Package: \tab coefficientalpha\cr
Type: \tab Package\cr
License: \tab GPL\cr
}
}
\author{
Zhiyong Zhang and Ke-Hai Yuan
Maintainer: Zhiyong Zhang <zzhang4@nd.edu>
}
\references{
Zhang, Z. & Yuan, K.-H. (2014). Robust Coefficients Alpha and Omega and their Confidence Intervals: Methods and Software.
}
\keyword{robust}
\keyword{alpha}
\keyword{Cronbach}

