\name{coefficientalpha-package}
\alias{coefficientalpha-package}
\alias{coefficientalpha}
\docType{package}
\title{
 Robust Cronbach's alpha and McDonald's omega with missing and non-normal data
}
\description{
An R package to calculate coefficient alpha and omega with missing data and non-normal data. Robust standard errors and confidence intervals are also provided.
}
\details{
\tabular{ll}{
Package: \tab coefficientalpha\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2013-01-22\cr
License: \tab GPL\cr
}
}
\author{
Zhiyong Zhang and Ke-Hai Yuan
Maintainer: Zhiyong Zhang <zzhang4@nd.edu>
}
\references{
Zhang, Z. & Yuan, K.-H. (2014). Robust Coefficients Alpha and Omega and their Confidence Intervals: Methods and Software.
}
\keyword{robust}
\keyword{alpha}
\keyword{Cronbach}

